project restructure and add copyright
