+++ src/WebSocketServer/Events/BlockEvent.cs
@@ -8,13 +8,9 @@
 // Redistribution and use in source and binary forms with or without
 // modifications are permitted.
 
-using System;
-namespace Neo.Plugins.WebSocketServer;
+namespace Neo.Plugins.WebSocketServer.Events;
 
-public class WssException : Exception
+public class BlockEvent : WebSocketEvent
 {
-    public WssException(int code, string message) : base(message)
-    {
-        HResult = code;
-    }
+    public int Height { get; set; }
 }
