+++ src/WebSocketServer/Events/WebSocketEvent.cs
@@ -8,13 +8,11 @@
 // Redistribution and use in source and binary forms with or without
 // modifications are permitted.
 
-using System;
-namespace Neo.Plugins.WebSocketServer;
+using Neo.Json;
+namespace Neo.Plugins.WebSocketServer.Events;
 
-public class WssException : Exception
+public abstract class WebSocketEvent
 {
-    public WssException(int code, string message) : base(message)
-    {
-        HResult = code;
-    }
+    public WssEventId WssEvent { get; set; }
+    public JObject Data { get; set; }
 }
